CASMPET-3907: Switch sonar-jobs-watcher to use crictl to stop istio-proxy
The sonar-jobs-watcher CronJob was using `kubectl exec pkill` in the
istio-proxy container to stop it when it noticed that a Job was ready
to exit. This worked because the istio-proxy image had pkill in it.

We'd like to switch to a distroless image of the proxyv2 image and this
doesn't include pkill, so the old sonar-jobs-watcher script won't work
anymore.

The proposed change is to execute `crictl stop` to stop the istio-proxy
container instead. This requires

1) Switching to an image that has crictl in it (the new cary/sonar
   image)
2) Mounting the crictl config file and socket from the host into the
   sonar container so that crictl can affect the containers on the
   host.
3) Switch from a CronJob to a DaemonSet so that the script is running
   on each host.

The other changes here are related to switching from a CronJob that
starts a fresh pod every time to a script that runs a loop with a
delay, and having to get the istio-proxy container's ID so that it can
be passed to `crictl stop`.

A Dockerfile is added to build the image and the Jenksisfile is
modified to build the image.