CASMPET-5067: Clean up orphaned CFS pods (CSM-0.9)
On a customer system we were seeing some CFS Pods not get cleaned
up. It was expected that these Pods would be cleaned up by the
sonar-jobs-watcher. The sonar-jobs-watcher wasn't cleaning up
these pods because they didn't match the query when fetching the
list of Pods to clean up because they were orphaned. The
owning Job was gone and the ownerReferences weren't there for the
Pod anymore. It's unknown why the Pods were orphaned.

The proposed workaround is to change sonar-jobs-watcher to
look for the CFS pods. These Pods have a `cfsession` label.

This is a backport from the 1.0+ version. In 0.9 sonar-jobs-watcher
was still using a CronJob and used `kubectl exec` to stop the
sidecar.