CASMPET-5182: sonar-jobs-watcher restarts when containerd restarted
The sonar-jobs-watcher would sometimes hit a problem where the
attempt to kill the istio-proxy would fail and time out, and this
would make it so that jobs never finished. The proposed fix is to
have the sonar-jobs-watcher exit with an error when this happens
and then kubernetes will just restart it so that it can re-try with
a fresh connection to containerd.